ajout de la sécurité login sur les pages concernées, début de la page I/E

+++ views/deleteQuestion.php
@@ -1,32 +1,36 @@
 <?php
     require_once("../functions/fonctions.php");
-    // récupération de l'id de la question en GET
-    $id = $_GET['id'];
-    $idQ = $_GET['idQ'];
-    
-    // connexion à la base de données
-    require_once("../functions/fonctions.php");
-    $db = connect_db();
+    if (!verifierSiIlEstConnecte()) {
+        header('Location: ./home.php');
+    } else {
 
-    // suppression des réponses de la question
-    $reponse = "DELETE FROM REPONSE WHERE idQuestion = $id";
-    // exécution de la requête
-    $db->query($reponse);
+        // récupération de l'id de la question en GET
+        $id = $_GET['id'];
+        $idQ = $_GET['idQ'];
+        
+        // connexion à la base de données
+        require_once("../functions/fonctions.php");
+        $db = connect_db();
 
-    // suppression de la question
-    $question = "DELETE FROM QUESTION WHERE idQuestion = $id";
-    // exécution de la requête
-    $db->query($question);
-    
-    $questions = getQuestions($idQ);
-    if (count($questions) != 0) {
-        // On actualise les numéros des questions
-        miseAJourDesNumerosQuestions($idQ);
+        // suppression des réponses de la question
+        $reponse = "DELETE FROM REPONSE WHERE idQuestion = $id";
+        // exécution de la requête
+        $db->query($reponse);
+
+        // suppression de la question
+        $question = "DELETE FROM QUESTION WHERE idQuestion = $id";
+        // exécution de la requête
+        $db->query($question);
+        
+        $questions = getQuestions($idQ);
+        if (count($questions) != 0) {
+            // On actualise les numéros des questions
+            miseAJourDesNumerosQuestions($idQ);
+        }
+
+        
+
+        // redirection vers la page de listage des questionnaires
+        header("Location: editQuestionnaire.php?id=$idQ");
     }
-
-    
-
-    // redirection vers la page de listage des questionnaires
-    header("Location: editQuestionnaire.php?id=$idQ");
-
 ?>
